requirements/base.txt
django==3.1.6
django-environ==0.4.5
psycopg2-binary==2.8.6
django-braces==1.14.0
pypugjs==5.9.8
django-crispy-forms==1.11.0
django-cleanup==5.1.0
Markdown==3.3.3
djangorestframework==3.12.2
requests==2.25.1
Pillow==8.1.0
django-datatables-view==1.19.1
